fix: run agent script

+++ run_agent.sh
@@ -1,4 +1,8 @@
-rm -r tsunami
+if test -d tsunami; then
+  echo "Removing previous agent build"
+  rm -r tsunami
+fi
+
 find . -empty -type d -delete  # remove empty directories to avoid wrong hashes
 autonomy packages lock
 autonomy fetch --local --agent dvilela/tsunami
